+++ src/restSpa/config.nim
@@ -1,0 +1,2 @@
+const
+  sess_username* = "username"